Trying to use an auxiliar file to serve the app.server instance to gunicorn.
Now using `gunicorn production_app:wikichron` should be equivalent to what was in the past of doing `gunicorn app:server`.
However, it has the same issue as we had before: Two workers respond to one request, when they absolutely should not! :S
This would be a very nice way to keep using gunicorn and to have this wikichron module isolation. If it doesn't work the
 next step could be to check using uWSGI.